requirements.txt
python-dotenv==0.21.0
slack-sdk==3.18.3
